requirements.txt
torch
torchvision
pillow
pyyaml
opencv-python
numpy